Deluxe_Consignment/requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Django==3.1
django-ckeditor==6.0.0
django-filter==2.3.0
django-google-analytics-app==4.4.2
django-js-asset==1.2.2
django-materializecss-form==1.1.17
django-multiselectfield==0.1.12
django-storages==1.9.1
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
Pillow==7.2.0
postgres==3.0.0
protobuf==3.13.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.10.0
python-decouple==3.3
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10